fixed lossy conversions

+++ Assignments/ass2/test/euclidean_vector/Constructors/default_constructor.cpp
@@ -6,6 +6,6 @@
 
 TEST_CASE("default_constructor") {
     const auto vector = comp6771::euclidean_vector();
-    REQUIRE(vector.dimensions() == 1);
+    REQUIRE(static_cast<size_t>(vector.dimensions()) == 1);
     REQUIRE(vector[0] == 0);
 }
