+++ Assignments/ass2/test/euclidean_vector/Constructors/size_constructor.cpp
@@ -7,7 +7,7 @@
 
 TEST_CASE("size_constructor_implicit") {
     const auto vector = comp6771::euclidean_vector(10);
-    REQUIRE(vector.dimensions() == 10);
+    REQUIRE(static_cast<size_t>(vector.dimensions()) == 10);
     // check all values defaulting to 0
     REQUIRE(std::all_of(vector.begin(), vector.end(),
                         [&](auto value) { return value == 0; }));
@@ -16,12 +16,12 @@ TEST_CASE("size_constructor_implicit") {
 TEST_CASE("size_constructor_explicit") {
     auto dimension = int{3};
     auto vector_object_initialised = comp6771::euclidean_vector(dimension);
-    REQUIRE(vector_object_initialised.dimensions() == 3);
+    REQUIRE(static_cast<size_t>(vector_object_initialised.dimensions()) == 3);
     REQUIRE(std::all_of(vector_object_initialised.begin(), vector_object_initialised.end(),
                         [&](auto value) { return value == 0; }));
 }
 
 TEST_CASE("size_constructor_empty") {
     const auto vector = comp6771::euclidean_vector(0);
-    REQUIRE(vector.dimensions() == 0);
+    REQUIRE(static_cast<size_t>(vector.dimensions()) == 0);
 }
