+++ Assignments/ass2/test/euclidean_vector/Operations/unary.cpp
@@ -8,24 +8,24 @@
 
 TEST_CASE("unary_base_case") {
     auto vector = comp6771::euclidean_vector();
-    REQUIRE(vector.dimensions() == 1);
+    REQUIRE(static_cast<size_t>(vector.dimensions()) == 1);
     REQUIRE(vector[0] == 0);
     vector = +vector;
-    REQUIRE(vector.dimensions() == 1);
+    REQUIRE(static_cast<size_t>(vector.dimensions()) == 1);
     REQUIRE(vector[0] == 0);
 }
 
 TEST_CASE("unary_changes") {
     auto vector = comp6771::euclidean_vector();
-    REQUIRE(vector.dimensions() == 1);
+    REQUIRE(static_cast<size_t>(vector.dimensions()) == 1);
     REQUIRE(vector[0] == 0);
 
     vector = +vector;
-    REQUIRE(vector.dimensions() == 1);
+    REQUIRE(static_cast<size_t>(vector.dimensions()) == 1);
     REQUIRE(vector[0] == 0);
 
     // make sure changes give desired results
     vector[0] = 1;
-    REQUIRE(vector.dimensions() == 1);
+    REQUIRE(static_cast<size_t>(vector.dimensions()) == 1);
     REQUIRE(vector[0] == 1);
 }
